mixed precision training seems to work (but is non-deterministic right now)

+++ src/nn_models/resnet/imagenet_training_cuda.cu
@@ -53,7 +53,7 @@
 namespace rlt = RL_TOOLS_NAMESPACE_WRAPPER ::rl_tools;
 namespace fs = std::filesystem;
 
-using T = float;
+using T = __nv_bfloat16;
 using TYPE_POLICY = rlt::numeric_types::Policy<float,
     rlt::numeric_types::UseCase<rlt::numeric_types::categories::Parameter, T>,
     rlt::numeric_types::UseCase<rlt::numeric_types::categories::Activation, T>,
